add function for three cards
